requirements.txt
langchain
torch
accelerate
sentence_transformers
streamlit_chat
streamlit
faiss-cpu
huggingface-hub
pypdf
python-dotenv
docx2txt
langchain-community
langchain_groq